Filtered through hard coded data to get unique accounts.
The getBalance and toString methods in the customer class were updated to print the balance and change the output of toString respectively. In the Main class, the raw data was filtered through and each unique account was added to a customer list. Any duplicate ids in the customer list were then filtered through and their account data was added to their respective accounts. Then the positive and negative accounts were listed.